% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{add_quantile}
\alias{add_quantile}
\title{A dataset containing the cutoff value between negative and positive populations for each marker}
\format{
A data frame with 20 rows and 2 variables:
\describe{
  \item{filename}{ character: the name of the marker for each cutoff}
  \item{quantile_99}{ numeric: the corresponding MFI value which is the cutoff between positive and negative populations for a particular marker}
}
}
\usage{
add_quantile
}
\description{
A dataset containing the cutoff value between negative and positive populations for each marker
}
\keyword{datasets}
