% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_fe}
\alias{all_fe}
\title{A dataset containing the feature engineered data}
\format{
A dataframe with 37,143 rows and 22 variables:
\describe{
  \item{IL_10}{ factor: the positive (1) or negative (0) expression of IL-10 in a cell}
  \item{TNF}{ factor: the positive (1) or negative (0) expression of TNF in a cell}
  \item{CD4}{ factor: the positive (1) or negative (0) expression of CD4 in a cell}
  \item{CD8}{ factor: the positive (1) or negative (0) expression of CD8 in a cell}
  \item{filename}{ character: the sample name from which each data point originates}
  \item{cell_no}{ numeric: the number of cells with the particular phenotype listed in the markers}
  \item{total_count_by_file}{ numeric: the total number of cells in each sample}
  \item{percentage}{ numeric: the percentage of cells with a particular phenotype listed in the markers}
  ...
}
}
\usage{
all_fe
}
\description{
A dataset containing the feature engineered data
}
\keyword{datasets}
