% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.numeric.factor.R
\name{as.numeric.factor}
\alias{as.numeric.factor}
\title{Convert factor variables to numeric while retaining the numbers}
\usage{
\method{as.numeric}{factor}(x)
}
\arguments{
\item{x}{vector of factors variables}
}
\value{
vector of numeric variables
}
\description{
Convert factor variables to numeric while retaining the numbers
}
\examples{

x <- as.factor(0, 1)
as.numeric.factor(x)

}
