% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_corr.R
\name{calc_corr}
\alias{calc_corr}
\title{Calculate correlation matrix of cell percentages for individual populations}
\usage{
calc_corr(df)
}
\arguments{
\item{df}{dataframe that contains the samples, populations, and percentages}
}
\value{
correlation matrix
}
\description{
Calculate correlation matrix of cell percentages for individual populations
}
\examples{
calc_corr(sample_populations_all_groups)

}
