% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_flowjo.R
\name{check_flowjo}
\alias{check_flowjo}
\title{Compare calculated feature engineered percentages with Flowjo manual percentage}
\usage{
check_flowjo(df, marker)
}
\arguments{
\item{df}{dataframe that contains the samples, populations, and percentages}

\item{marker}{column name for the name of the marker that you want to check}
}
\value{

}
\description{
Compare calculated feature engineered percentages with Flowjo manual percentage
}
\examples{
check_flowjo(all_fe, CD3)

}
