% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_factor_numeric.R
\name{convert_factor_numeric}
\alias{convert_factor_numeric}
\title{Convert factor variables to numeric while retaining the numbers}
\usage{
convert_factor_numeric(x)
}
\arguments{
\item{x}{vector of factors variables}
}
\value{
vector of numeric variables
}
\description{
Convert factor variables to numeric while retaining the numbers
}
\examples{

x <- as.factor(c(0, 1))
convert_factor_numeric(x)


}
