% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_calc.R
\name{count_calc}
\alias{count_calc}
\title{Calculates the cell counts and percentages for individual populations}
\usage{
count_calc(df)
}
\arguments{
\item{df}{dataframe to calculate the cell counts and percentages}
}
\value{
dataframe with the percentage of cells in each population for each sample
}
\description{
Calculates the cell counts and percentages for individual populations
}
\examples{
df_all_gated_simple \%>\%
 select(ends_with("-A"), -`FSC-A`, filename) \%>\%
 dplyr::rename(`CD3` = "Alexa Fluor 532-A",
               `CD4` = "BV480-A",
               `CD8` = "BV570-A") \%>\%
 dplyr::filter(`SSC-A` != max(`SSC-A`)) \%>\%
 mutate(CD3 = fe(add_quantile, CD3, "CD3"),
        CD4 = fe(add_quantile, CD4, "CD4"),
        CD8 = fe(add_quantile, CD8, "CD8")) \%>\%
 select(-`Zombie Nir-A`,  -`SSC-A`) \%>\%
 count_calc()

}
