% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_FMO_gated_data}
\alias{df_FMO_gated_data}
\title{A dataframe containing all of the FMO samples and corresponding measured data}
\format{
A dataframe with 415,544 rows that represent each cell in the FMO samples and 24 variables which show the different markers containing all of the measured data:
\describe{
  \item{SSC-A}{ numeric: the sidescatter, or complexity, of each cell}
  \item{filename}{ character: the name of the FMO sample from which the data comes}
  \item{IL_10}{ numeric: the MFI expression of IL-10 for each cell}
  \item{TNF}{ numeric: the MFI expression of TNF for each cell}
  \item{CD4}{ numeric: the MFI expression of CD4 for each cell}
  \item{CD69}{ numeric: the MFI expression of CD69 for each cell}
  \item{CD8}{ numeric: the MFI expression of CD8 for each cell}
  ...
}
}
\usage{
df_FMO_gated_data
}
\description{
A dataframe containing all of the FMO samples and corresponding measured data
}
\keyword{datasets}
