% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe.R
\name{fe}
\alias{fe}
\title{Feature engineer the data}
\usage{
fe(quantile_df, marker, marker_name)
}
\arguments{
\item{quantile_df}{the dataframe that contains the list of cutoffs for each marker}

\item{marker}{name of the marker column to feature engineer}

\item{marker_name}{name of the marker column to feature engineer in quotes}
}
\value{
convert the marker data column into either a 0 for negative expression or 1 for positive expressoin
}
\description{
Feature engineer the data
}
\examples{
df_all_gated \%>\%
rename("CD3" = "Alexa Fluor 532-A") \%>\%
mutate(CD3 = fe(add_quantile, CD3, "CD3")

}
