% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_for_total_pheno.R
\name{filter_for_total_pheno}
\alias{filter_for_total_pheno}
\title{Filter for visualizing all the phenotypes}
\usage{
filter_for_total_pheno(df, marker_order)
}
\arguments{
\item{df}{dataframe for which you want to visualize the total number of different populations}

\item{marker_order}{vector of the order you want the markers to be in for the plot}
}
\value{

}
\description{
Filter for visualizing all the phenotypes
}
\examples{
order_of_markers <- c("CD3", "CD4", "CD8",  "CD44", "CD103", "Sca1", "IL_17","CTLA4", "CD27",  "CD153", "KLRG1", "IFN",  "CD122", "PD1", "CD62L", "IL_10", "CD28","TNF")
filter_for_total_pheno(all_fe, marker_order = order_of_markers)

#' @importFrom dplyr \%>\%

}
