% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pops.R
\name{filter_pops}
\alias{filter_pops}
\title{Filter populations to show all the populations of interest}
\usage{
filter_pops(df)
}
\arguments{
\item{df}{dataframe that contains the information for all of the populations}
}
\value{
dataframe that contains the filtered population names
}
\description{
Filter populations to show all the populations of interest
}
\examples{
library(dplyr)
all_fe \%>\%
 dplyr::filter(CD3 == 1 & percentage > 0.5) \%>\%
 filter_pops()

}
