% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowset_gated_data}
\alias{flowset_gated_data}
\title{A FlowSet containing all of the data samples and measured data}
\format{
A flowSet with 18 experiments (samples) containing all of the measured data:
\describe{
  \item{SSC-A}{ numeric: the sidescatter, or complexity, of each cell}
  \item{filename}{ character: the name of the FMO sample from which the data comes}
  \item{IL_10}{ numeric: the MFI expression of IL-10 for each cell}
  \item{TNF}{ numeric: the MFI expression of TNF for each cell}
  \item{CD4}{ numeric: the MFI expression of CD4 for each cell}
  \item{CD69}{ numeric: the MFI expression of CD69 for each cell}
  \item{CD8}{ numeric: the MFI expression of CD8 for each cell}
  ...
}
}
\usage{
flowset_gated_data
}
\description{
A FlowSet containing all of the data samples and measured data
}
\keyword{datasets}
