% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FMO_filtered_data}
\alias{FMO_filtered_data}
\title{A dataset containing the matched FMO markers to the FMO samples}
\format{
A data frame with 415,544 rows and 4 variables:
\describe{
  \item{SSC-A}{ numeric: the sidescatter used to plot the FMO data}
  \item{filename}{ character: the name of the FMO sample from which the data comes}
  \item{marker}{ character: the name of the marker of interest that matches up with the name of the FMO sample}
  \item{MFI}{ numeric: mean fluorescent intensity data for each FMO sample in the specific marker channel}
  ...
}
}
\usage{
FMO_filtered_data
}
\description{
A dataset containing the matched FMO markers to the FMO samples
}
\keyword{datasets}
