% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_corr.R
\name{format_corr}
\alias{format_corr}
\title{Format correlation matrix for input into ggplot}
\usage{
format_corr(corr_matrix)
}
\arguments{
\item{corr_matrix}{a correlation matrix derived from using stats::cor()}
}
\value{
convert matrix to tidy data with columns for `Var1` and `Var2` (the populations) and `value` (the correlation value)
}
\description{
Format correlation matrix for input into ggplot
}
\examples{
corr <- calc_corr(sample_populations_all_groups)
format_corr(corr)

}
