% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_picked_pops.R
\name{heatmap_picked_pops}
\alias{heatmap_picked_pops}
\title{Plot heatmap of the selected populations of interest}
\usage{
heatmap_picked_pops(df, order_of_markers)
}
\arguments{
\item{df}{dataframe that contains the populations (listed as 0's and 1's), the population numbers, and the phenotype descriptions}

\item{order_of_markers}{vector of the order of markers to display on the plot x-axis}
}
\value{
heatmap separated by lineages
}
\description{
Plot heatmap of the selected populations of interest
}
\examples{
heatmap_picked_pops(phenotype_data, order_of_markers)

}
