% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_subset_pheno.R
\name{heatmap_subset_pheno}
\alias{heatmap_subset_pheno}
\title{Visualize subset of identified populations in a heatmap}
\usage{
heatmap_subset_pheno(df, marker_order)
}
\arguments{
\item{df}{dataframe for which you want to visualize a small subset of populations}
}
\value{
pheatmap of the smaller subset of phenotypes denoted by a combination of positive and negative expression of markers
}
\description{
Visualize subset of identified populations in a heatmap
}
\examples{
library(dplyr)

 order_of_markers <- c("CD3", "CD4", "CD8",  "CD44", "CD103", "Sca1", "IL_17","CTLA4",
"CD27",  "CD153", "KLRG1", "IFN",  "CD122", "PD1", "CD62L", "IL_10", "CD28","TNF")

heatmap_subset_pheno(simple_pop_df, order_of_markers)

#' @importFrom dplyr \%>\%

}
