% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FMOs.R
\name{plot_FMOs}
\alias{plot_FMOs}
\title{Title}
\usage{
plot_FMOs(df1, df2)
}
\arguments{
\item{df1}{the dataframe that contains the filtered FMO values (only the FMO file that matches up with it's marker)}

\item{df2}{the dataframe that contains the quantile values}
}
\value{

}
\description{
Title
}
