% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_populations}
\alias{sample_populations}
\title{A dataset containing the filtered population data}
\format{
A dataframe with 33 rows and 19 variables:
\describe{
  \item{IL_10}{ factor: the positive (1) or negative (0) expression of IL-10 in a cell}
  \item{TNF}{ factor: the positive (1) or negative (0) expression of TNF in a cell}
  \item{CD4}{ factor: the positive (1) or negative (0) expression of CD4 in a cell}
  \item{CD8}{ factor: the positive (1) or negative (0) expression of CD8 in a cell}
  \item{population}{ character: the population name of the filtered data}
  ...
}
}
\usage{
sample_populations
}
\description{
A dataset containing the filtered population data
}
\keyword{datasets}
