% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_populations_all_groups}
\alias{sample_populations_all_groups}
\title{A dataset containing the percentages for each population and sample in the data}
\format{
A dataframe with 594 rows and 3 variables:
\describe{
  \item{population}{ character: the population name of the filtered data}
  \item{filename}{ character: the sample name from which each data point originates}
  \item{percentage}{ numeric: the percentage of cells with a particular phenotype listed in the markers}
  ...
}
}
\usage{
sample_populations_all_groups
}
\description{
A dataset containing the percentages for each population and sample in the data
}
\keyword{datasets}
