% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_flow_set.R
\name{tidy_flow_set}
\alias{tidy_flow_set}
\title{Apply the tidy_flow_item function to a flowSet}
\usage{
tidy_flow_set(flowSet)
}
\arguments{
\item{flowSet}{the flowset in which to convert to a tidy flow set}
}
\value{
Returns a dataframe with a row for each observation (cell) with columns for:
\itemize{
item{\code{filename}}: {the name of the sample for which the cell comes from}
item{\code{fluorescent-markers}}: {the same number of marker-fluorescent pairs on read
on the machine will show as columns in this data frame}
}
}
\description{
This function ultimately converts a flowSet into a tidy data frame
}
\examples{
tidy_flow_set(flowset_FMO_gated_data)

}
