% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_calc.R
\name{count_calc}
\alias{count_calc}
\title{Calculates the cell counts and percentages for individual populations}
\usage{
count_calc(df)
}
\arguments{
\item{df}{dataframe to calculate the cell counts and percentages}
}
\value{
dataframe with the percentage of cells in each population for each sample
}
\description{
Calculates the cell counts and percentages for individual populations
}
