% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_all_gated}
\alias{df_all_gated}
\title{A dataset containing all of the sample data}
\format{
A data frame with 1,056,969 rows and 28 variables:
\describe{
  \item{filename}{ character: the sample name from which each data point originates}
  ...
  }
}
\usage{
df_all_gated
}
\description{
A dataset containing all of the sample data
}
\keyword{datasets}
