% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FMO_filtered_data}
\alias{FMO_filtered_data}
\title{A dataset containing all of the filtered FMO data for use in calculating the}
\format{
A dataframe with 415,544 rows that represent each cell in the FMO samples and 4 variables which show corresponding filename, marker, and MFI values:
\describe{
  \item{SSC-A}{ numeric: the MFI expression for the Side Scatter (complexity) Area for each cell}
  \item{filename}{ character: the name of the FMO sample from which the data comes}
  \item{marker}{ character: the name of the marker for which MFI values are saved}
  \item{MFI}{ numeric: the MFI expression for the corresponding marker in the "marker" column}
  ...
}
}
\usage{
FMO_filtered_data
}
\description{
A dataset containing all of the filtered FMO data for use in calculating the
}
\keyword{datasets}
