% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_99.R
\name{get_99}
\alias{get_99}
\title{Title}
\usage{
get_99(df)
}
\arguments{
\item{df}{the dataframe that contains the filtered FMO (so only the FMO filename has data for the FMO marker)}
}
\value{

}
\description{
Title
}
\examples{

}
