% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identified_pop_perc.R
\name{identified_pop_perc}
\alias{identified_pop_perc}
\title{Add in the percentage data for all of the samples for the filtered populations}
\usage{
identified_pop_perc(df1, df2, marker_vector)
}
\arguments{
\item{df1}{dataframe that contains the filtered populations}

\item{df2}{dataframe that contains the feature engineered data with the percentages}

\item{marker_vector}{vector of the order of markers to show on the plots}
}
\value{
dataframe that contains percentage of cells in each of the filtered populations
}
\description{
Add in the percentage data for all of the samples for the filtered populations
}
\examples{
identified_pop_perc(sample_populations, all_fe, marker_vector = order_of_markers)

}
