% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FMOs.R
\name{plot_FMOs}
\alias{plot_FMOs}
\title{Plot all of the FMOs and their cutoffs}
\usage{
plot_FMOs(df1, df2)
}
\arguments{
\item{df1}{the dataframe that contains the filtered FMO values (only the FMO file that matches up with it's marker)}

\item{df2}{the dataframe that contains the quantile values}
}
\value{
plots that shows data for each of the FMOs and where the cutoff is placed
}
\description{
Plot all of the FMOs and their cutoffs
}
\examples{
plot_FMOs(FMO_filtered_data, add_quantile)

#' @importFrom rlang .data

}
