% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_flow_item.R
\name{tidy_flow_item}
\alias{tidy_flow_item}
\title{Convert flowframe to tidy dataframe}
\usage{
tidy_flow_item(flowSet_item)
}
\arguments{
\item{flowSet_item}{the flowset in which to convert to a tidy flow set}
}
\value{

}
\description{
This function is used by tidy_flow_set
}
\examples{

}
